requirements.txt
snowflake-snowpark-python
nbconvert
pytest